HW6/requirements.txt
python-dotenv
openai>=1.40.0
transformers>=4.44.0
torch
